Added more fake canary types and made them more configurable
