+++ mingming-controller/manifest.yml
@@ -5,5 +5,3 @@ applications:
   memory: 1G
   host: mingming
   timeout: 120
-  env:
-    FAKEWORKERS: true
