+++ mingming-controller/src/main/java/net/guidowb/mingming/MingMingController.java
@@ -1,7 +1,5 @@
 package net.guidowb.mingming;
 
-import java.net.URI;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
@@ -9,7 +7,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.core.env.Environment;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
-import net.guidowb.mingming.controllers.FakeWorkers;
+import net.guidowb.mingming.fakes.FakeWorkers;
 
 @SpringBootApplication
 @EnableScheduling
@@ -23,9 +21,6 @@ public class MingMingController implements CommandLineRunner {
 
 	@Override
 	public void run(String... args) throws Exception {
-	    String serverPort = env.getProperty("SERVER_PORT", "8080");
-	    String serverHost = env.getProperty("vcap.application.uris[0]", "localhost:" + serverPort);
-	    URI serverURI = URI.create("http://" + serverHost);
-		if (env.containsProperty("FAKEWORKERS")) FakeWorkers.create(serverURI, 10);
+		FakeWorkers.start(env);
 	}
 }
